+++ tools/qfcc/source/dot_expr.c
@@ -466,7 +466,7 @@ _print_expr (dstring_t *dstr, expr_t *e, int level, int id, expr_t *next)
 		return;
 	e->printid = id;
 
-	if (e->type < 0 || e->type > ex_value) {
+	if ((int) e->type < 0 || e->type > ex_value) {
 		dasprintf (dstr, "%*se_%p [label=\"(bad expr type)\\n%d\"];\n",
 				   indent, "", e, e->line);
 		return;
